patch: Remove PHY LED / enable GPIO P2.1
This patch removes the "phy_led" devices from the EA3250 board specific
code. As far as I can see, this is a leftover from the code specific to
the PHY board, as there is no LED connected to P2.1.

Moreover, the joystick switch (left) on the QVGA Base Board is connected
to P2.1, so without this patch, it will not work.

Patch applies to linux-2.6.39.2-lpc
  git://git.lpclinux.com/linux-2.6.39.2-lpc
  126a5edc8e66550c5d7bed4209283f41cf4a8159