patch: Setup P2 as GPIO
This patch sets the P2 MUX state to GPIO in the EA3250 board setup code.
Before this, P2 was set up as EMC_D[31:19] (the default), which is unused
on this board as the DDR is connected through a 16 bit databus (according
to the documentation).

Hence, this patch effectively enables one to use the 12 GPIOs of P2
in the generic GPIO driver.

Patch applies to linux-2.6.39.2-lpc
  git://git.lpclinux.com/linux-2.6.39.2-lpc
  126a5edc8e66550c5d7bed4209283f41cf4a8159