patch: SPD_CUST handling for HSUART driver
This patch adds support for SPD_CUST handling to HSUART driver, so it
will be possible to set non-standard baud rates using the TIOCGSERIAL /
TIOCSSERIAL ioctls.

One caveat is that the generic serial driver code in the kernel exports
the "baud_base" in the TIOCGSERIAL ioctl as uart_clk/16, while uart_clk/14
would be a better fit for the HSUARTs.  So userspace needs to take this
into account when setting the custom divisor.  (i.e., calculate it as:
`custom_divisor = (13000000 + 7 * baudrate) / (14 * baudrate) - 1`)

Patch applies to linux-2.6.39.2-lpc
  git://git.lpclinux.com/linux-2.6.39.2-lpc
  126a5edc8e66550c5d7bed4209283f41cf4a8159